Moved duplicate code to utils, minimized secrets fetch
