+++ slack_bot/bundle_reader.sh
@@ -5,5 +5,5 @@ pip install --platform manylinux2014_x86_64 --implementation cp --only-binary=:a
 
 #remove extraneous bits from installed packages
 rm -r dist_reader/*.dist-info
-cp config.py models.py message_reader.py dist_reader/
+cp config.py models.py utils.py message_reader.py dist_reader/
 cd dist_reader && zip -r lambda.zip *
