+++ slack_bot/bundle_writer.sh
@@ -7,5 +7,5 @@ pip install -I openai -t dist_writer
 
 #remove extraneous bits from installed packages
 rm -r dist_reader/*.dist-info
-cp config.py models.py chain.py message_writer.py dist_writer/
+cp config.py models.py utils.py chain.py message_writer.py dist_writer/
 cd dist_writer && zip -r lambda.zip *
